+++ CHANGELOG.md
@@ -7,3 +7,18 @@
 * Initial Commit
     - List PIM users
     - Delete PIM user
+    
+    
+1.0.0-beta2
+-------
+
+* Added new PIM user commands
+    - Change Password PIM user
+    - Create PIM user
+    
+    
+1.0.0-beta3
+-------
+
+* Added media commands
+    - Images - Remove Orphans